Extract MovieEqualityComparer to separate file.
